Refactor tests and security enhancements
