+++ lib/agents.ts
@@ -131,7 +131,7 @@ Additional rules:
 - DISTILL the research into polished README content. Strip all researcher artifacts: "File references:", "What I found:", source-file citations, notes-to-self, and any raw analysis scaffolding. The reader should see only clean documentation, never the research process.
 - Do NOT repeat information across sections. Each fact belongs in exactly one section — place it where the template guidance says it goes.
 - Your entire output must be ONLY the raw README markdown — no preamble, no closing commentary, no wrapping code fences`,
-    tools: [readFile],
+    tools: [],
     handoffs: [detailFetcher],
     handoffDescription:
       'Write the final README using accumulated context',
@@ -155,7 +155,7 @@ Additional rules:
 - Omit any section where nothing concrete was found
 - If a required section is missing specific facts, hand off to DetailFetcher to retrieve it. Limit yourself to 3 handoffs.
 - Your entire output must be ONLY the raw AGENTS.md markdown — no preamble, no closing commentary, no wrapping code fences`,
-        tools: [readFile],
+        tools: [],
         handoffs: [detailFetcher],
         handoffDescription: 'Write the AGENTS.md using accumulated context',
       })
