+++ package.json
@@ -28,7 +28,7 @@
     "dev": "tsx ./script.ts",
     "build": "tsc",
     "start": "node script.js",
-    "test": "jest ./script.spec.ts",
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest",
     "check": "tsx ./script.ts --check",
     "refresh": "tsx ./script.ts",
     "refresh:interactive": "tsx ./script.ts --interactive",
@@ -43,6 +43,7 @@
   "dependencies": {
     "@clack/prompts": "1.0.1",
     "@openai/agents": "0.4.10",
+    "globby": "16.1.1",
     "markdownlint-cli": "0.47.0",
     "picocolors": "1.1.1",
     "typescript": "5.9.3",
